requirements.txt
flask>=0.10.0
flask-sqlalchemy>=2.0.0
flask-script>=2.0.5,<3
flask-migrate>=1.6.0
requests>=2.7.0
psycopg2>=2.6.1,<3
alembic==0.8.3
pytest-cov>=2.0.0,<3
flake8>=2.5.1
webtest>=2.0.18,<3
